move archive tests to qc and cleanup
